Organize auth endpoint serializers
This just moves the serializers into a serializers.py file and gives them an inheritance hierarchy.
